refactor: Organizating folders
