+++ Curso_Video/backend/db/fake_db.py
@@ -1,0 +1,3 @@
+from typing import List
+
+db: List[dict] = []
