+++ Curso_Video/backend/middleware/cors_config.py
@@ -1,0 +1,10 @@
+from fastapi.middleware.cors import CORSMiddleware
+
+def add_cors_middleware(app):
+    app.add_middleware(
+        CORSMiddleware,
+        allow_origins=["*"], 
+        allow_credentials=True,
+        allow_methods=["*"],  
+        allow_headers=["*"],  
+    )
