+++ Curso_Video/backend/models/CourseModel.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+from typing import Optional
+
+class CourseModel(BaseModel):
+    id: str | None = None
+    name: str
+    price: float
